Added version 2.1 of the bioreactor firmware
Unpack the zip file, open the contained *.ino file in the Arduino IDE and compile. The firmware needs to run on a genuine Arduino Nano 33 IoT. The genuine part is important as Arduino clones often do not come with an individual serial number. This serial number is important for keeping track of several bioreactors connected to one control PC.